Get catalog from table func context
