+++ crates/catalog/Cargo.toml
@@ -11,7 +11,6 @@ logutil = { path = "../logutil" }
 object_store = { workspace = true }
 parking_lot = "0.12.1"
 protogen = { path = "../protogen" }
-sqlbuiltins = { path = "../sqlbuiltins" }
 thiserror.workspace = true
 tokio = { workspace = true }
 tonic = { workspace = true }
