+++ crates/datafusion_ext/Cargo.toml
@@ -13,6 +13,7 @@ unicode_expressions = []
 [dependencies]
 ioutil = { path = "../ioutil" }
 telemetry = { path = "../telemetry" }
+catalog = { path = "../catalog" }
 serde_json = { workspace = true }
 datafusion = { workspace = true }
 async-trait = "0.1.74"
