+++ crates/sqlbuiltins/src/functions/virtual_listing.rs
@@ -204,12 +204,12 @@ async fn get_db_lister(
     ctx: &dyn TableFuncContextProvider,
     dbname: String,
 ) -> Result<Box<dyn VirtualLister>> {
-    let db = ctx
-        .get_database_entry(&dbname)
-        .ok_or(ExtensionError::MissingObject {
+    let db = ctx.get_session_catalog().resolve_database(&dbname).ok_or(
+        ExtensionError::MissingObject {
             obj_typ: "database",
             name: dbname,
-        })?;
+        },
+    )?;
 
     let lister: Box<dyn VirtualLister> = match &db.options {
         DatabaseOptions::Internal(_) => {
