+++ crates/sqlexec/src/dispatch/external.rs
@@ -11,7 +11,9 @@ use datafusion::datasource::file_format::FileFormat;
 use datafusion::datasource::TableProvider;
 use datafusion::execution::context::SessionState;
 use datafusion::prelude::SessionContext;
-use datafusion_ext::functions::{FuncParamValue, TableFuncContextProvider, VirtualLister};
+use datafusion_ext::functions::{
+    DefaultTableContextProvider, FuncParamValue, TableFuncContextProvider, VirtualLister,
+};
 use datafusion_ext::vars::SessionVars;
 use datasources::bigquery::{BigQueryAccessor, BigQueryTableAccess};
 use datasources::common::url::DatasourceUrl;
@@ -47,7 +49,6 @@ use sqlbuiltins::functions::BUILTIN_TABLE_FUNCS;
 
 use catalog::session_catalog::SessionCatalog;
 
-use super::listing::CatalogLister;
 use super::{DispatchError, Result};
 
 /// Dispatch to external tables and databases.
@@ -57,30 +58,6 @@ pub struct ExternalDispatcher<'a> {
     df_ctx: &'a SessionContext,
     /// Whether or not local file system access should be disabled.
     disable_local_fs_access: bool,
-}
-
-impl<'a> TableFuncContextProvider for ExternalDispatcher<'a> {
-    fn get_database_entry(&self, name: &str) -> Option<&DatabaseEntry> {
-        self.catalog.resolve_database(name)
-    }
-
-    fn get_credentials_entry(&self, name: &str) -> Option<&CredentialsEntry> {
-        self.catalog.resolve_credentials(name)
-    }
-
-    fn get_session_vars(&self) -> SessionVars {
-        let cfg = self.df_ctx.copied_config();
-        let vars = cfg.options().extensions.get::<SessionVars>().unwrap();
-        vars.clone()
-    }
-
-    fn get_session_state(&self) -> SessionState {
-        self.df_ctx.state()
-    }
-
-    fn get_catalog_lister(&self) -> Box<dyn VirtualLister> {
-        Box::new(CatalogLister::new(self.catalog.clone(), true))
-    }
 }
 
 impl<'a> ExternalDispatcher<'a> {
@@ -546,7 +523,11 @@ impl<'a> ExternalDispatcher<'a> {
         };
         let prov = resolve_func
             .unwrap()
-            .create_provider(self, args, opts)
+            .create_provider(
+                &DefaultTableContextProvider::new(self.catalog, self.df_ctx),
+                args,
+                opts,
+            )
             .await?;
         Ok(prov)
     }
